requirements.txt
flask
line-bot-sdk
beautifulsoup4
requests
python-dotenv
